CHORE : 거리 계산 구현체 누락건 추가

+++ src/main/java/com/tripcut/domain/location/repository/LocationReviewRepository.java
@@ -20,5 +20,5 @@ public interface LocationReviewRepository extends JpaRepository<LocationReview, 
     Double getAverageRatingByLocationId(@Param("locationId") Long locationId);
     
     @Query("SELECT COUNT(A) FROM LocationReview A WHERE A.filmingLocation.id = :locationId")
-    Integer getReviewCountByLocationId(@Param("locationId") Long locationId);
+    Double getReviewCountByLocationId(@Param("locationId") Long locationId);
 }
